Create separate script for populating database
